TaskTracker/tasktracker/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
channels
